Renames layout component file
